+++ src/Sensor/PressureSensor.h
@@ -18,11 +18,11 @@ public:
 	PressureSensor(I2C &i2c);
 	virtual ~PressureSensor() {};
 	float toHz();
-	int toValue();
+	float toValue();
 private:
 	uint8_t pressureData[3];
 	uint8_t readPressureCmd ;
-	int16_t pressure;
+	float pressure;
 	I2C& i2c;
 };
